requirements/docs.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=./requirements/docs.txt --pre ./requirements.in/docs.txt
#
alabaster==0.7.12         # via sphinx
argh==0.26.2              # via sphinx-autobuild
babel==2.8.0              # via sphinx
certifi==2019.11.28       # via requests
chardet==3.0.4            # via doc8, requests
doc8==0.8.0
docutils==0.16            # via doc8, flit, restructuredtext-lint, sphinx
flit-core==2.2.0          # via flit
flit==2.2.0               # via sphinx-copybutton
idna==2.8                 # via requests
imagesize==1.2.0          # via sphinx
jinja2==3.0.0a1           # via sphinx
livereload==2.6.1         # via sphinx-autobuild
markupsafe==1.1.1         # via jinja2
packaging==20.1           # via sphinx
pathtools==0.1.2          # via sphinx-autobuild, watchdog
pbr==5.4.4                # via sphinx-click, stevedore
port_for==0.3.1           # via sphinx-autobuild
pydocstyle==5.0.2
pygments==2.5.2           # via sphinx
pyparsing==2.4.6          # via packaging
pytoml==0.1.21            # via flit, flit-core
pytz==2019.3              # via babel
pyyaml==5.3               # via sphinx-autobuild
requests==2.22.0          # via flit, sphinx
restructuredtext-lint==1.3.0  # via doc8
six==1.14.0               # via doc8, livereload, packaging, stevedore
snowballstemmer==2.0.0    # via pydocstyle, sphinx
sphinx-autobuild==0.7.1
sphinx-autodoc-typehints==1.10.3
sphinx-click==2.3.1
sphinx-copybutton==0.2.8
sphinx-rtd-theme==0.4.3
sphinx==2.4.1
sphinxcontrib-applehelp==1.0.1  # via sphinx
sphinxcontrib-devhelp==1.0.1  # via sphinx
sphinxcontrib-htmlhelp==1.0.2  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.2  # via sphinx
sphinxcontrib-serializinghtml==1.1.3  # via sphinx
stevedore==1.32.0         # via doc8
tornado==6.0.3            # via livereload, sphinx-autobuild
urllib3==1.25.8           # via requests
watchdog==0.10.2          # via sphinx-autobuild
wheel==0.34.2             # via sphinx-copybutton

# The following packages are considered to be unsafe in a requirements file:
# setuptools
